move collectionlog
