+++ src/commands/Minion/championchallenge.ts
@@ -1,8 +1,8 @@
 import { CommandStore, KlasaMessage } from 'klasa';
 
 import { BotCommand } from '../../lib/BotCommand';
-import { championScrolls } from '../../lib/collectionLog';
 import { Activity, Time } from '../../lib/constants';
+import { championScrolls } from '../../lib/data/collectionLog';
 import { MinigameIDsEnum } from '../../lib/minions/data/minigames';
 import { minionNotBusy, requiresMinion } from '../../lib/minions/decorators';
 import { UserSettings } from '../../lib/settings/types/UserSettings';
