+++ src/commands/Minion/equippet.ts
@@ -2,7 +2,7 @@ import { CommandStore, KlasaMessage } from 'klasa';
 import { Item } from 'oldschooljs/dist/meta/types';
 
 import { BotCommand } from '../../lib/BotCommand';
-import { pets } from '../../lib/collectionLog';
+import { pets } from '../../lib/data/collectionLog';
 import { requiresMinion } from '../../lib/minions/decorators';
 import { UserSettings } from '../../lib/settings/types/UserSettings';
 import { removeItemFromBank } from '../../lib/util';
