+++ src/commands/OSRS_Fun/raid.ts
@@ -4,7 +4,7 @@ import { ItemBank } from 'oldschooljs/dist/meta/types';
 import ChambersOfXeric from 'oldschooljs/dist/simulation/minigames/ChambersOfXeric';
 
 import { BotCommand } from '../../lib/BotCommand';
-import { coxLog } from '../../lib/collectionLog';
+import { coxLog } from '../../lib/data/collectionLog';
 import { addBanks } from '../../lib/util';
 import createReadableItemListFromBank from '../../lib/util/createReadableItemListFromTuple';
 import filterBankFromArrayOfItems from '../../lib/util/filterBankFromArrayOfItems';
