+++ src/lib/data/eatables.ts
@@ -1,5 +1,5 @@
-import { Eatable } from './skilling/types';
-import itemID from './util/itemID';
+import { Eatable } from '../skilling/types';
+import itemID from '../util/itemID';
 
 export const Eatables: Eatable[] = [
 	{
