+++ src/lib/minions/data/killableMonsters/bosses/gwd.ts
@@ -1,7 +1,7 @@
 import { Monsters } from 'oldschooljs';
 
-import { bosses } from '../../../../collectionLog';
 import { Time } from '../../../../constants';
+import { bosses } from '../../../../data/collectionLog';
 import itemID from '../../../../util/itemID';
 import resolveItems, { deepResolveItems } from '../../../../util/resolveItems';
 import { KillableMonster } from '../../../types';
